requirements.txt
# Core AI & Agent Framework
anthropic>=0.40.0
langgraph>=0.2.0
langchain-core>=0.3.0

# UI
streamlit>=1.40.0

# Data
pandas>=2.0.0

# Environment
python-dotenv>=1.0.0

# Dev & testing
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.8.0
black>=24.0.0
mypy>=1.10.0
